Added additional validation to validate_tx_hash function if transaction was today. Refactored code base
